Update README.md
# AI-INTERNAL-AFFAIRS
My "Building AI course" project
An attempt to create an AI tool to fight against bias in media articles, social network posts, web pages descriptions....

## Summary

We all have some unconscious gender biases which we can not notice when we speak or write up. This also happens to journalists, media people or influenceres who 
inadvertently contribute to perpetuate these prejudices. To break out this harmful chain, the idea is to create an AI tool they can use to revise their texts before posting them in order to identify these unconscious biases.

## Background

In our daily life we can find instances of unconscious male chauvinism such as the football coach shouting "stop running like a girl" to those players who are not making the most,or a newspaper article about the last World Economic Forum describing females outfits but not their males partners. Most of times those are unconscious but they are harmful anyways. They are contributing to perpetuate gender inequalty.To help us to avoid those, I'd like to create an AI tool.


## How is it used?

It works as a single grammar checker or writing assistant.The app Grammarly (helping non English speakers to write properly in English by checking the grammar of the texts written) is a good reference. It can be added as a new keyboard in your smartphone to make it all even easier.
It will be useful for anyone writing up and posting articles, master cases, end of degree works... or even novels. In other words, whoever wanting to be readen.

# write your solution here
Internal Affairs (IA) pretends to be an AI tool that uses machine learning to help anyone to avoid unconscious gender biases.

Using Natural Language Processing and neural network for classification will classify the texts in gender balanced or unbalanced, and will propose corrections for the last one ones.

The model will be trained with data sets from accredited authors and organizations such as UN Women, Women Faculty Forum (yale university), or some other relevant institutions in this area in different countries and languages

## Challenges

The main problem is that expanding across languages When it is time to expand to another countries with differente languages, we´ll need to seek a solution on how to utilize our existing model in a new market because the tool has been trained with Spanish or English or Deustch.. data and as such it could not be applied to incoming cases in another different language. 

## What next?

To make it able to work in any language. I've been reading aboputlanguage-agnostic word embeddings and I guess they could be a good way to explore through.


## Acknowledgments

“Building AI course project” by University of Helsinki
Reaktor
https://somebuddy.com/